requirements.txt
flask
flask-sqlalchemy
Flask-WTF
flask-login
werkzeug
pandas
pymysql
redis
python-dotenv
MarkupSafe
itsdangerous
suds-jurko
